Share the instrumentation exclusion policy between the two instrumenters
isClassnameInstrumented, is_transformer and junitTestClasses each existed
twice, once per instrumenter.  An exclusion added to only one copy would
silently change comparability depending on which instrumenter ran.  Premain
already owns most of what the policy consults and compiles on every JDK.

The one real difference between the copies, the set_class_initialized rule,
is kept and commented: it can only fire under DCInstrument24, because
DCInstrument adds that call to the already-instrumented <clinit>.
